Improve the bytecode optimizer's ability to simplify `letrec's  to smaller `letrec' groups or even `let*'.  The goal of the change is to avoid performance surprises  when using internal definitions, especially when mixing  experessions with definitions.  It's a somewhat scary change. Besides the new transformation,  the optimizer pass's coordinate system for `letrec' (and sometimes  `let*') bindings changed.  No standard benchmarks were harmed during the making of this  optimization. (None were improved, either.)

+++ src/racket/src/port.c
@@ -7436,7 +7436,7 @@ static void unused_process_group(void *_sp, void *ignored)
 # else
   void **unused_group;
   unused_group = malloc(sizeof(void *) * 2);
-  unused_group[0] = (void *)sp->pid;
+  unused_group[0] = (void *)(long)sp->pid;
   unused_group[1] = unused_groups;
   need_to_check_children = 1;
 # endif
